Moving Message Componento to Common folder
