--- src/app/investor/message/message.component.html
@@ -1,75 +1,0 @@
-<!-- ///////////////////////////////////////////////////////////////////////////////// -->
-<!-- //////////////////////////   MAIN Content  ////////////////////////////////////// -->
-<!-- ///////////////////////////////////////////////////////////////////////////////// -->
-  <div>
-    <div class="row"> <!--  Page name, Search Bar -->
-      <div class="col-xs-4 col-sm-4 col-md-4">
-        <p class="sw-page-name">Transaction confirmation</p>
-      </div>
-    </div> <!-- END Page name, Search Bar  -->
-  
-    <div class="row">
-      <div class="col-xs-2 col-xs-offset-3 col-sm-2 col-sm-offset-3 col-md-2 col-md-offset-3">
-        <p class="sw-title-outside-box sw-border-dark-red">Transaction confirmation</p>
-      </div>
-      <div class="col-xs-3 col-sm-3 col-md-3">
-        <p class="sw-description-outside-box" [ngClass]="{'sw-box-title': true, 'sw-error-message': messageService.getError()}">{{messageService.getHeaderMessage() || 'Confirm transaction'}}</p>
-      </div>
-      <div class="col-xs-2 col-sm-2 col-md-2 col-xs-offset-1 col-sm-offset-1 col-md-offset-1">
-        <p class="sw-button-add-funds sw-btn-add-funds" role="button" [routerLink]="['/investor/add-funds/child']">Add funds</p>
-      </div>
-    </div>
-
-    <div class="row"> <!-- //////////// MAIN Content //////////// -->
-      <div class="col-sm-7 col-sm-offset-1 col-md-7 col-md-offset-1">
-
-          <div class=" col-md-12"> <!-- Start  -->
-            <div class="sw-box">
-              <div class="row sw-box-space">
-
-              <div class="row">
-                <div class="col-sm-1 col-md-1">
-                  <img src="assets/img/swapy-logo4.png" alt="Swapy logo" class="swapy-logo">
-                </div>
-                <div class="col-sm-10 col-md-10">
-                    <p class="sw-box-title" [ngClass]="{'sw-box-title': true, 'sw-error-message': messageService.getError()}" *ngIf="!messageService.getMessage()">You have not submitted your transaction yet.<br>Please, search for the Metamask wallet icon (<img src="assets/img/icon-64.png" class="metamask-logo">) at the top right section of your window and click it to proceed.</p>
-
-                    <p class="sw-box-title" [ngClass]="{'sw-box-title': true, 'sw-error-message': messageService.getError()}" *ngIf="messageService.getMessage()">
-                      {{messageService.getMessage()}}
-                    </p>
-                </div>
-              </div>
-
-              <div class="row">
-                <div class="col-sm-4 col-md-4 col-sm-offset-1 col-md-offset-1">
-                  <img *ngIf="messageService.getLoadingState()" src="assets/img/grid.svg" alt="Loading logo" class="loading">
-                </div>
-                <div class=" col-sm-7 col-md-7" *ngIf='messageService.getError() === false'>
-                    <p class="sw-box-text sw-blue1">Your wallet address: {{walletAddress}}</p>
-                </div>
-              </div>
-
-                <div class="row">
-                  <div class=" col-sm-5 col-md-5 col-sm-offset-3 col-md-offset-3"  *ngIf='messageService.getError() === false'>
-                    <p class="sw-box-text sw-successful-message">Now you can follow your investment status on: <span routerLink="/investor" class="sw-successful-bold" role="button">Manage</span></p>
-                  </div>
-                  <div class=" col-sm-5 col-md-5 col-sm-offset-3 col-md-offset-3"  *ngIf='messageService.getError() === true'>
-                    <p class="sw-box-text sw-successful-message">You can go back and do the transaction again</p>
-                  </div>
-                  <div class=" col-sm-3 col-md-3"*ngIf='messageService.getError() === false'>
-                    <a class="sw-button sw-btn-ghost1" routerLink="/investor">Close</a>
-                  </div>
-                  <div class=" col-sm-3 col-md-3" *ngIf='messageService.getError() === true'>
-                    <a class="sw-button sw-btn-ghost1" routerLink="/investor">Close</a>
-                  </div>
-                </div>
-
-
-               </div>
-            </div>
-
-          </div>
-        </div> <!-- END  -->
-
-      </div>
-    </div><!-- END MAIN Content  -->
